feat(terraform): add example .tfvars file for Terraform configuration
This commit adds a new file `.example_terraform.tfvars` to the `terraform` directory. This file serves as an example for configuring Terraform variables. It includes the following variables:

- `billing_account`: Your billing account ID.
- `folder_id`: The ID of the folder where the project will be created.
- `credential_path`: The path to your service account key file.

This example file can be used as a template for creating the actual `.terraform.tfvars` file with the appropriate values for your specific setup.